+++ src/core/entities/Simulation.ts
