+++ src/core/entities/Worker.ts
